test(int): split type and behavior tests
