+++ test/no-content.test.ts
@@ -1,6 +1,6 @@
 import { HttpResponse, type StrictResponse } from "msw";
 import { createOpenApiHttp } from "openapi-msw";
-import { describe, expect, expectTypeOf, test } from "vitest";
+import { describe, expect, test } from "vitest";
 import type { paths } from "./fixtures/no-content.api.js";
 
 describe("Given an OpenAPI schema endpoint with no-content", () => {
@@ -32,22 +32,5 @@ describe("Given an OpenAPI schema endpoint with no-content", () => {
 
     expect(result?.response?.body).toBeNull();
     expect(result?.response?.status).toBe(201);
-  });
-
-  test("When an endpoint is mocked, Then responses with content cannot be returned", async () => {
-    type Endpoint = typeof http.delete<"/resource">;
-    const resolver = expectTypeOf<Endpoint>().parameter(1);
-    const response = resolver.returns.extract<Response>();
-
-    response.not.toEqualTypeOf<StrictResponse<{ id: number }>>();
-  });
-
-  test("When an endpoint is mocked, Then responses must be strict responses", async () => {
-    type Endpoint = typeof http.delete<"/resource">;
-    const resolver = expectTypeOf<Endpoint>().parameter(1);
-    const response = resolver.returns.extract<Response>();
-
-    response.not.toEqualTypeOf<Response>();
-    response.toEqualTypeOf<StrictResponse<null>>();
   });
 });
