+++ test/path-fragments.test.ts
@@ -1,6 +1,6 @@
 import { HttpResponse } from "msw";
 import { createOpenApiHttp } from "openapi-msw";
-import { describe, expect, expectTypeOf, test } from "vitest";
+import { describe, expect, test } from "vitest";
 import type { paths } from "./fixtures/path-fragments.api.js";
 
 describe("Given an OpenAPI schema endpoint that contains path fragments", () => {
@@ -19,22 +19,6 @@ describe("Given an OpenAPI schema endpoint that contains path fragments", () => 
     const responseBody = await result?.response?.json();
     expect(responseBody?.id).toBe("test-id");
     expect(responseBody?.name).toBe("test-name");
-  });
-
-  test("When an endpoint is mocked, Then the path fragments are strict-typed", () => {
-    type Endpoint = typeof http.get<"/resource/{id}/{name}">;
-    const resolver = expectTypeOf<Endpoint>().parameter(1);
-    const params = resolver.parameter(0).toHaveProperty("params");
-
-    params.toEqualTypeOf<{ id: string; name: string }>();
-  });
-
-  test("When a endpoint contains no path fragments, Then no params are provided", () => {
-    type Endpoint = typeof http.get<"/resource">;
-    const resolver = expectTypeOf<Endpoint>().parameter(1);
-    const params = resolver.parameter(0).toHaveProperty("params");
-
-    params.toEqualTypeOf<never>();
   });
 
   // FIXME: See https://github.com/christoph-fricke/openapi-msw/issues/22
